Añadidos multiples modelos y endpoints
Añadidos modelos y funciones restantes a la API

+++ src/main/java/com/example/SportyRest/repository/ActividadRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
@@ -18,14 +19,17 @@ public interface ActividadRepository extends JpaRepository<Actividad, Integer> {
     void deleteById(int id);
 
     // Devuelve las actvidades dado el id del
-    List<Actividad> findByCreador(int creador);
+    List<Actividad> findByCreador(Long creador);
 
+    //List<Actividad> findByFechaHoraFinalizacionBefore(LocalDateTime ahora);
 
     @Query("SELECT a FROM Actividad a WHERE " + "earth_distance(ll_to_earth(:latitud, :longitud)," +
             " ll_to_earth(a.latitud, a.longitud)) <= :distancia")
     List<Actividad> findActividadesCercanas(@Param("latitud") double latitud,
                                             @Param("longitud") double longitud,
                                             @Param("distancia") double distancia);
+
+
 
 
 }
